Break the view module into multiple smaller modules.
