+++ frontend-cli/src/services/api.ts
@@ -1,11 +1,5 @@
 import * as rp from 'request-promise';
-import { GameId, GameplayCommandName, TableState } from '../types';
-
-export interface Game {
-  gameId: string;
-  table: TableState;
-  score: number;
-}
+import { Game, GameId, GameplayCommandName } from '../types';
 
 const json = true;
