+++ frontend-cli/src/views/default/footer.ts
@@ -1,0 +1,14 @@
+import { Card } from '../../types';
+import { top } from '../../util';
+import { margin } from './strings';
+import { generateCardView } from './card';
+
+export const generateFooterView = (stock: Card[], waste: Card[], tableauColumns: number) => [
+  '',
+  [
+    ...Array.from({ length: tableauColumns - 2 }, () => generateCardView(undefined, false)),
+    generateCardView(top(waste)),
+    generateCardView(top(stock))
+  ].join(margin),
+  ''
+];
