+++ frontend-cli/src/views/default/foundation.ts
@@ -1,0 +1,16 @@
+import { Card } from '../../types';
+import { top } from '../../util';
+import { baseSlot, margin } from './strings';
+import { generateCardView } from './card';
+
+export const generateFoundationView = (foundation: Card[][], tableauColumns: number) => [
+  '',
+  [
+    ...Array.from(
+      { length: tableauColumns - foundation.length },
+      () => baseSlot
+    ),
+    ...foundation.map(top).map((card) => generateCardView(card))
+  ].join(margin),
+  ''
+];
